requirements.txt
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2
django-dotenv==1.4.2
django-heroku==0.3.1
djangorestframework==3.12.4
gunicorn==20.1.0
psycopg2==2.9.1
pytz==2020.5
requests==2.25.1
requests-html==0.10.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.4.3
whitenoise==5.3.0